refactor(shorebird_cli): move executable wrapper classes to src/executables directory
