+++ packages/shorebird_cli/lib/src/executables/adb.dart
