+++ packages/shorebird_cli/lib/src/executables/git.dart
