+++ packages/shorebird_cli/lib/src/executables/java.dart
