+++ packages/shorebird_cli/lib/src/executables/xcodebuild.dart
